Move `IconGroup` out of `...Windows.Native` namespace
It was really bogus that `SDL2DesktopWindow` was importing `osu.Framework.Platform.Windows.Native`.
